refactor: 이미지, 고지서 서비스 분리
