+++ server/src/main/java/com/oss/carbonadministrator/domain/gas/GasInfo.java
@@ -1,4 +1,4 @@
-package com.oss.carbonadministrator.domain.water;
+package com.oss.carbonadministrator.domain.gas;
 
 import com.oss.carbonadministrator.domain.base.BaseTimeEntity;
 import javax.persistence.Column;
@@ -14,51 +14,34 @@ import lombok.NoArgsConstructor;
 /*
  * TODO 인식 데이터 변경
  */
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "water")
-public class WaterInfo extends BaseTimeEntity {
+@Table(name = "gas")
+public class GasInfo extends BaseTimeEntity {
 
     @Id
-    @Column(name = "water_id", nullable = false)
+    @Column(name = "gas_id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    // 기본 요금
     @Column
     private int demandCharge;
 
-    // 연료비조정액
-    @Column
-    private int fuelAdjustmentRate;
-
-    // 부가가치세
     @Column
     private int vat;
 
-    // 월단위절사
     @Column
     private int roundDown;
 
-    // 당월요금계
     @Column
     private int totalbyCurrMonth;
 
     @Column
     private int totalPrice;
 
-    // 당월 사용량 (kWh)
     @Column
     private int currMonthUsage;
-
-    // 전월 사용량 (kWh)
-    @Column
-    private int preMonthUsage;
-
-    // 전년동월 사용량 (kWh)
-    @Column
-    private int lastYearUsage;
-
 }
